chore(movie details): renaming components
