+++ src/screens/MovieDetails/MovieDetails.types.ts
