+++ src/screens/MovieDetails/index.ts
@@ -1,1 +1,1 @@
-export { MovieDetails } from './screen';
+export { MovieDetails } from './MovieDetails';
