--- src/screens/MovieDetails/screen/index.ts
@@ -1,1 +1,0 @@
-export { MovieDetails } from './Screen';
